Fixed CommentRegularExpression rule so that only one instance is created
Previously a new instance was created for each instance of the rule, whereas
now only one instance is created that contains multiple rules.

The config setting name has had to be changed so that 'Key' is always the key
name for consistency with all other rules (instead of 'ParentKey') and the
individual instance of the rule is now named 'ChildKey' instead of 'Key'.
This makes the parsing of the config file entirely generic instead of needing
a special case just for this rule.