Make cbde binaries available in unit tests (#2658)
Make sure that even when a very light build is done (for instance, from the IDE), the executable files from CBDE are registered as "content" in the SonarAnalyzer.CSharp.csproj project. In order to do that, we add them twice:
- Once statically, when they are already here, this addition is always executed
- Once dynamically, when they are just downloaded, this addition may be skipped if the solution is pretty much up-to-date
Then add those file with a more classical path relative to the project.