Fix crashes; and more tests (#2962)
* Improve display of exception
* Ignore C# 8 syntaxes
* Prevent crash in ThreadCpuStopWatch when running in .NET core
In .NET core, the AppDomain.GetCurrentThreadId() PAI does not return the physical thread id, but something else. As a consequence, we cannot get the processor time in this environment, and we will return -1 instead.

* Replace JSON by XML in the result file from CBDE
All libraries that we have seen to parse JSON are different depending on which platform they run on (.NET core vs .NET framework), and are therefore not suitable for use in our strict environment which requires one assembly compatible with all run-time environments. As a consequence, we changed the output format from JSON to XML, since XML can be parsed more easily and more consistently.