Fix OpenCover branch coverage import by using BranchPoints
When a code file is touched by unit tests from different test projects,
it will appear multiple times in the open cover reports. Since different
tests can cover different branches we need to correctly identify the
covered branches and count them only once.

SequencePoint does not have enough information to allow this
and because of that we have to rely on BranchPoints.