Allow import of internal issues with invalid location
Fixes #2484

Issues from third-party Roslyn analyzers (including Roslyn analyzers
provided by Microsoft) are included in the MSBuild output and imported
by default into SonarQube.

In this case they are considered external issues and the roslyn importer
is very forgiving regarding the location:
- if the column is invalid it will use only the line
- if both column and line are invalid, it will report the issue on the file.

If, the SDK for SonarQube Roslyn Analyzer Plugins is used to create
a plugin for the analyzers, then the issues are treated as internal and the
importer will not fallback to line or file level.

This commit is changing that and allows the import of internal issues
with invalid location.