Reuse NullPointerCheck in SE rules
NullPointerCheck is added by default by the CSharpExplodedGraph in order
to stop the visit early when a null pointer dereference is found.

This check is later overwritten by PublicMethodArgumentsShouldBeCheckedForNull
and NullPointerDereference.

In order to be able to run all the rules within one single semantic
execution pass we have to reuse the check.