Add support for ITestBuilder attributes in NUnit.
NUnit3 added support for custom test attributes via the ITestBuilder
interface. The analyzer for S2187 would not consider test methods
decorated with such attributes, giving a false positive result if the
fixture only contained such methods.

Because this interface only exists in NUnit3, tests for it are split
into a separate file, so that analyzer tests can be executed only for
NUnit version 3.0.0 and up.