Revert BUILD-3696 Use agent pool on AWS
Temporary revert until the assembly signing is fixed.
```
smctl sync:
smctl : The term 'smctl' is not recognized as the name of a cmdlet, function, script file, or operable program. Check
the spelling of the name, or if a path was included, verify that the path is correct and try again.
```